claude refactor
