+++ apps/src/sketchlab/excalidraw/styles/sketchlab-view.module.scss
