+++ apps/src/sketchlab/excalidraw/utils/generateNewExternalFiles.ts
