+++ apps/src/sketchlab/excalidraw/utils/populateInitialExcalidrawState.ts
