+++ apps/src/sketchlab/excalidraw/utils/uploadExternalFiles.ts
